Update to Blazor-State 0.4.6
ResetState and redirect
Code Cleanup
Remove Idology
