+++ Source/Herc.Pwa.Client/Features/Application/ApplicationState.Behavior.cs
@@ -1,0 +1,18 @@
+﻿namespace Herc.Pwa.Client.Features.Application
+{
+  using BlazorState;
+
+  public partial class ApplicationState : State<ApplicationState>
+  {
+    public ApplicationState() { }
+
+    protected ApplicationState(ApplicationState aState) : this()
+    {
+      Name = aState.Name;
+    }
+
+    public override object Clone() => new ApplicationState(this);
+
+    protected override void Initialize() => Name = "wallet.herc.one";
+  }
+}
