+++ Source/Herc.Pwa.Client/Features/Application/ApplicationState.cs
@@ -1,25 +1,8 @@
 ﻿namespace Herc.Pwa.Client.Features.Application
 {
-  using BlazorState;
-
-  public partial class ApplicationState : State<ApplicationState>
+  public partial class ApplicationState 
   {
-    public ApplicationState() { }
-
-    protected ApplicationState(ApplicationState aState) : this()
-    {
-      Name = aState.Name;
-    }
-
     public string Name { get; private set; }
-    public string Version { get; set; }
-
-    public override object Clone() => new ApplicationState(this);
-
-    protected override void Initialize()
-    {
-      Name = "Herc Progressive Web Application";
-      Version = "0.0.1";
-    }
+    public string Version => GetType().Assembly.GetName().Version.ToString();
   }
 }
