+++ Source/Herc.Pwa.Client/Features/Edge/EdgeCurrencyWallet/EdgeCurrencyWalletsState.cs
@@ -1,34 +1,10 @@
 ﻿namespace Herc.Pwa.Client.Features.Edge.EdgeCurrencyWallet
 {
-  using System;
   using System.Collections.Generic;
   using System.Linq;
-  using System.Reflection;
-  using BlazorState;
 
-  public partial class EdgeCurrencyWalletsState : State<EdgeCurrencyWalletsState>
+  public partial class EdgeCurrencyWalletsState
   {
-    public EdgeCurrencyWalletsState()
-    {
-      EdgeCurrencyWallets = new Dictionary<string, EdgeCurrencyWallet>();
-    }
-    protected EdgeCurrencyWalletsState(EdgeCurrencyWalletsState aEdgeCurrencyWalletsState) : this()
-    {
-      Console.WriteLine("EdgeCurrencyWalletState");
-      
-      aEdgeCurrencyWalletsState
-        .EdgeCurrencyWallets
-        .ToList()
-        .ForEach
-        (
-          aKeyValuePair =>
-            EdgeCurrencyWallets[aKeyValuePair.Key] =
-              aKeyValuePair.Value.Clone() as EdgeCurrencyWallet
-        );
-
-      SelectedEdgeCurrencyWalletId = aEdgeCurrencyWalletsState.SelectedEdgeCurrencyWalletId;
-    }
-
     public Dictionary<string, EdgeCurrencyWallet> EdgeCurrencyWallets { get; set; }
 
     public string SelectedEdgeCurrencyWalletId { get; set; }
@@ -48,22 +24,6 @@
       }
     }
     
-    public override object Clone() => new EdgeCurrencyWalletsState(this);
-
-    protected override void Initialize()
-    {
-      // Default will be an empty List
-    }
-
-    /// <summary>
-    /// Use in Tests ONLY, to initialize the State
-    /// </summary>
-    /// <param name="aCount"></param>
-    //public void Initialize()
-    //{
-    //  ThrowIfNotTestAssembly(Assembly.GetCallingAssembly());
-    //  Count = aCount;
-    //}
   }
 
 }
