--- Source/Herc.Pwa.Client/Pages/Idology.cshtml
@@ -1,9 +1,0 @@
-﻿@page "/idology"
-@layout Herc.Pwa.Client.Layouts.HercLayout.HercLayout
-@inherits IdologyModel
-
-<h1>@ApplicationState.Name</h1>
-
-<div>Idology Page</div>
-
-<style></style>
